bpo-38806: bdb: stop_here: always stop at calling frame
The following will not stop for debugging:

    python3.8 -c 'import pdb; pdb.Pdb(skip=["*"]).set_trace()'

The example is contrived, the real case would be to have some "noisy"
module being excluded in general, but when you add an explicit
"set_trace()" in there it should still stop there, and not on some upper
frame.

This was changed a long time already in
https://github.com/python/cpython/commit/313a7513b0c5771042d850d70782a2448d1cdcb7
(Python 2.3), but it is not really clear.

This PR is meant to see how reverting that part goes.